move all image contexts in a single file
