--- drawBot/context/bmpContext.py
@@ -1,8 +1,0 @@
-from __future__ import absolute_import
-
-from .imageContext import ImageContext
-
-
-class BMPContext(ImageContext):
-
-    fileExtensions = ["bmp"]
